Added new modules, including preparation, orca, postprocessing, and check. Code overhaul finished and distributed to these modules. Unit tests also updated.
